+++ packages/app-builder-lib/src/targets/win/winAppUtil.ts
@@ -1,18 +1,18 @@
 import { Arch, asArray, copyOrLinkFile, ensureNotBusy, escapeForXml, InvalidConfigurationError, log, sanitizeDirPath, walk } from "builder-util"
-import type { MsixWindowsService } from "../options/MsixOptions.js"
+import type { MsixWindowsService } from "../../options/MsixOptions.js"
 import { Nullish } from "builder-util-runtime"
 import _fsExtra from "fs-extra"
 const { emptyDir, readdir, readFile, writeFile } = _fsExtra
 import * as path from "path"
-import type { Target, TargetSpecificOptions } from "../core.js"
-import { FileAssociation } from "../options/FileAssociation.js"
-import { Protocol } from "../options/PlatformSpecificBuildOptions.js"
-import { getWindowsKitsBundle } from "../toolsets/winCodeSign.js"
-import { getTemplatePath } from "../util/pathManager.js"
-import { VmManager } from "../vm/vm.js"
-import type { WinPackager } from "../winPackager.js"
-import { createStageDir } from "./targetUtil.js"
-import { CAPABILITIES, isValidCapabilityName } from "./win/AppxCapabilities.js"
+import type { Target, TargetSpecificOptions } from "../../core.js"
+import { FileAssociation } from "../../options/FileAssociation.js"
+import { Protocol } from "../../options/PlatformSpecificBuildOptions.js"
+import { getWindowsKitsBundle } from "../../toolsets/winCodeSign.js"
+import { getTemplatePath } from "../../util/pathManager.js"
+import { VmManager } from "../../vm/vm.js"
+import type { WinPackager } from "../../winPackager.js"
+import { createStageDir } from "../targetUtil.js"
+import { CAPABILITIES, isValidCapabilityName } from "./AppxCapabilities.js"
 
 export const APPX_ASSETS_DIR_NAME = "appx"
