+++ packages/app-builder-lib/src/winPackager.ts
@@ -15,7 +15,7 @@ import { DIR_TARGET, Platform, Target } from "./core.js"
 import { isWindowsSigningDisabled, RequestedExecutionLevel, resolveWindowsSigningConfiguration, WindowsConfiguration } from "./options/winOptions.js"
 import { Packager } from "./packager.js"
 import { chooseNotNull, PlatformPackager } from "./platformPackager.js"
-import MsixTarget from "./targets/MsixTarget.js"
+import MsixTarget from "./targets/win/MsixTarget.js"
 import AppXTarget from "./targets/win/AppxTarget.js"
 import MsiTarget from "./targets/win/MsiTarget.js"
 import MsiWrappedTarget from "./targets/win/MsiWrappedTarget.js"
